requirements.txt
async-timeout==4.0.3
bottle==0.12.25
bottle-session==1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
idna==3.7
joblib==1.4.2
msal==1.28.0
mysql-connector-python==8.4.0
nltk==3.8.1
numpy==1.24.4
pandas==2.0.3
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
regex==2024.4.28
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.5.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.1
